pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yuantus-plm"
version = "0.1.0"
description = "YuantusPLM (元图PLM) core service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jsonschema>=4.20.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.9",
  "requests>=2.31.0",
  "sqlalchemy>=2.0.0",
  "typer>=0.12.0",
  "uvicorn[standard]>=0.30.0",
  # Private deployment dependencies
  "psycopg[binary]>=3.1.0",
  "alembic>=1.13.0",
  "boto3>=1.34.0",
]

[project.scripts]
yuantus = "yuantus.cli:main"
plm = "yuantus.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
